requirements.txt
Flask
pillow
flask_cors
numpy
requests
